+++ test/fixtures/rippled/index.js
@@ -16,6 +16,7 @@ module.exports = {
   },
   empty: require('./empty'),
   subscribe: require('./subscribe'),
+  subscribe_error: require('./subscribe_error'),
   unsubscribe: require('./unsubscribe'),
   account_objects: {
     normal: require('./account-objects'),
@@ -36,6 +37,10 @@ module.exports = {
     fabric: require('./book-offers'),
     usd_xrp: require('./book-offers-usd-xrp'),
     xrp_usd: require('./book-offers-xrp-usd')
+  },
+  ledger_data: {
+    first_page: require('./ledger-data-first-page'),
+    last_page: require('./ledger-data-last-page')
   },
   ledger_entry: {
     error: require('./ledger-entry-error')